change nilai spk and tkdn percentage data type

+++ app/GraphQL/Mutations/Pengadaan/CreatePengadaanMutation.php
@@ -91,7 +91,7 @@ class CreatePengadaanMutation extends Mutation
                 'description' => 'The job performer'
             ],
             'nilai_spk' => [
-                'type' => Type::int(),
+                'type' => Type::float(),
                 'description' => 'The SPK value'
             ],
             'anggaran' => [
@@ -103,7 +103,7 @@ class CreatePengadaanMutation extends Mutation
                 'description' => 'The HPS'
             ],
             'tkdn_percentage' => [
-                'type' => Type::int(),
+                'type' => Type::float(),
                 'description' => 'The TKDN percentage'
             ],
             'catatan' => [
